Changes the seed data.  Added Resolver to get all serviceCategories
